+++ app/src/main/kotlin/com/d4rk/androidtutorials/ui/android/webview/tabs/WebViewTabLayoutFragment.kt
@@ -1,0 +1,36 @@
+package com.d4rk.androidtutorials.ui.android.webview.tabs
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.d4rk.androidtutorials.R
+import com.d4rk.androidtutorials.databinding.FragmentWebviewLayoutBinding
+import me.zhanghai.android.fastscroll.FastScrollerBuilder
+import java.io.ByteArrayOutputStream
+import java.io.IOException
+import java.io.InputStream
+class WebViewTabLayoutFragment : Fragment() {
+    private lateinit var _binding: FragmentWebviewLayoutBinding
+    private val binding get() = _binding
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
+        _binding = FragmentWebviewLayoutBinding.inflate(inflater, container, false)
+        FastScrollerBuilder(binding.alertDialogScrollView).useMd2Style().build()
+        val inputStream: InputStream = resources.openRawResource(R.raw.text_webview_xml)
+        val byteArrayOutputStream = ByteArrayOutputStream()
+        var i: Int
+        try {
+            i = inputStream.read()
+            while (i != - 1) {
+                byteArrayOutputStream.write(i)
+                i = inputStream.read()
+            }
+            inputStream.close()
+        } catch (e: IOException) {
+            e.printStackTrace()
+        }
+        binding.webViewXMLText.text = byteArrayOutputStream.toString()
+        return binding.root
+    }
+}
+